requirements.txt
dill==0.3.6
numpy==1.21.6
pandas==1.3.5
transformers==4.25.1
torch==1.12.1
torchvision==0.13.1
tensorboard==2.11.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
pytorch-crf==0.7.2